requirements/requirements.txt
Django==3.2.16
django-bootstrap-v5==1.0.11
django-markdownx==3.0.1
django-tagulous==1.3.3
factory-boy==3.2.1
Faker==8.12.0
python-dotenv==0.21.0
nbformat==5.7.0
nbconvert==7.2.2
Jinja2==3.0.1
django-createsuperuserwithpassword>=2.0.0
django-debug-toolbar>=3.7